% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fit_rpagp_cpp}
\alias{fit_rpagp_cpp}
\title{The C++ backend for the RPAGP model sampler.}
\usage{
fit_rpagp_cpp(y, n_iter, theta0, hyperparam, seed = 0L)
}
\arguments{
\item{y}{A matrix of observed trial data.}

\item{n_iter}{The total number of MCMC iterations.}

\item{theta0}{A list of initial parameter values.}

\item{hyperparam}{A list of hyperparameter values.}

\item{seed}{An integer seed for reproducibility.}
}
\description{
This is the core C++ function that runs the MCMC sampler.
It is not intended to be called directly by the user. Please use the
R wrapper function \code{fit_rpagp()} instead.
}
\keyword{internal}
